requirements.txt
python-telegram-bot==12.8
python-dotenv==1.1.1
astral==3.2 